Refactor BnCKFwdInference::GetSolution for NHWC. Rename and move RunCKSolution to implicitgemm_ck_util.hpp
